GH-24 Post-predicate dice filtering implementation
Unfortunately, subset-style checks are not supported in SwiftData
predicates. The alternative with the lightest code overhead is simply to
filter the card list post-query.

Post-query filtering does not require the field to be a `@Relationship`,
so much of the other work done to create the Die type in the database is
largely unnecessary. I will consider reverting some of it for simplicity
and maintainability.